+++ src/app/app.module.ts
@@ -7,23 +7,26 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { CookieService } from 'ngx-cookie-service';
 import { environment } from 'src/environments/environment';
 import { AppRoutingModule } from './app-routing.module';
-
-// Component
-import { HttpMockInterceptor } from './commons/interceptors/http-mock.interceptor';
+import { HttpHerokuInterceptor } from './commons/interceptors/http-heroku.interceptor';
 import { HeaderComponent } from './components/header/header.component';
 import { FooterComponent } from './components/footer/footer.component';
-
-// Page
 import { AppComponent } from './app.component';
 import { SentenceCasePipe } from './pipes/sentence-case.pipe';
+import { HttpWgnInterceptor } from './commons/interceptors/http-wgn.interceptor';
 
-const MOCK_INTERCEPTOR_PROVIDER: Provider = {
+const MOCK_HEROKU_INTERCEPTOR_PROVIDER: Provider = {
     provide: HTTP_INTERCEPTORS,
-    useExisting: forwardRef(() => HttpMockInterceptor),
+    useExisting: forwardRef(() => HttpHerokuInterceptor),
     multi: true,
 };
 
-let mockProviders = [HttpMockInterceptor, MOCK_INTERCEPTOR_PROVIDER];
+const MOCK_WGN_INTERCEPTOR_PROVIDER: Provider = {
+    provide: HTTP_INTERCEPTORS,
+    useExisting: forwardRef(() => HttpWgnInterceptor),
+    multi: true,
+};
+
+let mockProviders = [HttpHerokuInterceptor, MOCK_HEROKU_INTERCEPTOR_PROVIDER, HttpWgnInterceptor, MOCK_WGN_INTERCEPTOR_PROVIDER];
 
 if (environment.production) {
     mockProviders = [];
